Refactor: 사용하지 않는 페이지 제거 및 전체 페이지 경로 수정
