+++ src/pages/sinitto/hello-call-list/components/call-request/index.ts
@@ -1,0 +1,1 @@
+export { default as CallRequest } from './CallRequest';
