+++ src/pages/sinitto/hello-call-list/components/index.ts
@@ -1,0 +1,1 @@
+export * from './call-request';
