+++ src/pages/sinitto/hello-call-report/components/report-detail/ReportDetail.tsx
