+++ src/pages/sinitto/hello-call-service/components/service-detail/index.ts
