--- src/pages/sinitto/hello-call/hello-call-list/components/index.ts
@@ -1,1 +1,0 @@
-export * from './request-list';
