--- src/pages/sinitto/hello-call/hello-call-list/components/request-list/index.ts
@@ -1,1 +1,0 @@
-export { default as RequestList } from './RequestList';
