--- src/pages/sinitto/hello-call/index.ts
@@ -1,5 +1,0 @@
-export { default as HelloCallPage } from './HelloCallPage';
-
-export * from './hello-call-list';
-export * from './hello-call-report';
-export * from './hello-call-service';
