+++ src/pages/sinitto/index.ts
@@ -1,4 +1,6 @@
 export * from './call-back';
 export * from './guide-line';
-export * from './hello-call';
+export * from './hello-call-list';
+export * from './hello-call-report';
+export * from './hello-call-service';
 export * from './mypage';
